Update 10.1. Calculator.py

+++ Tests/Tests 10-19/10.1 Calculator/10.1. Calculator.py
@@ -17,14 +17,14 @@ operations = {
 }
 def calculator():
     print(logo)
-    num1 = int(input("What's the first number?:\n"))
+    num1 = float(input("What's the first number?:\n"))
     for symbol in operations:
         print(symbol)
     should_continue = True
 
     while should_continue:
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("What's the next number?:"))
+        num2 = float(input("What's the next number?:"))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1, num2)
